feat(QiunetUtils): idGenerator 使用long类型

+++ QiunetUtils/src/main/java/org/qiunet/utils/id/DefaultIdGenerator.java
@@ -1,6 +1,6 @@
 package org.qiunet.utils.id;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /***
  *
@@ -9,18 +9,18 @@ import java.util.concurrent.atomic.AtomicInteger;
  **/
 public class DefaultIdGenerator implements IdGenerator {
 
-	private AtomicInteger generator;
+	private AtomicLong generator;
 
 	public DefaultIdGenerator() {
 		this(0);
 	}
 
-	public DefaultIdGenerator(int initVal) {
-		this.generator = new AtomicInteger(initVal);
+	public DefaultIdGenerator(long initVal) {
+		this.generator = new AtomicLong(initVal);
 	}
 
 	@Override
-	public int makeId() {
+	public long makeId() {
 		return this.generator.incrementAndGet();
 	}
 }
